refactor(BEB-23): move recursos compartilhados para o diretorio shared
