+++ src/app.module.ts
@@ -1,8 +1,8 @@
 import { Logger, Module } from '@nestjs/common';
 
 import { DatabaseModule } from './database/database.module';
-import { AuthModule } from './modules/common/auth/auth.module';
 import { UserModule } from './modules/users/user.module';
+import { AuthModule } from './shared/auth/auth.module';
 
 @Module({
   imports: [DatabaseModule, UserModule, AuthModule],
