+++ src/modules/users/user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { BcryptService } from 'src/shared/bcrypt/bcrypt.service';
 
-import { AuthService } from '../common/auth/services/auth.service';
+import { AuthService } from '../../shared/auth/services/auth.service';
 import { AuthUserController, CreateUserController } from './controllers';
 import { UserEntity } from './entities';
 import { UserRepository } from './repositories';
