+++ src/shared/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 
-import { UserModule } from '../../../modules/users/user.module';
+import { UserModule } from '../../modules/users/user.module';
 import { jwtConstants } from '../constants/jwt.constants';
 import { AuthService } from './services/auth.service';
