+++ src/shared/entities/base-entity.ts
